chore : added changed requested in github, removed float32array

+++ src/x/__tests__/xAdd.test.ts
@@ -16,7 +16,7 @@ describe('xAdd', () => {
 
   it('test xAdd of array and floatarray', () => {
     let array1 = [10, 11, 12, 13, 14];
-    let array2 = new Float32Array([5, 4, 3, 2, 1]);
+    let array2 = new Float64Array([5, 4, 3, 2, 1]);
     expect(Array.from(xAdd(array1, array2))).toStrictEqual([
       15, 15, 15, 15, 15,
     ]);
